Work on issues
#9, #10, #11, #23, #26, #28, #30

+++ src/GCell/mainwindow/datawindow/cdatatable.h
@@ -16,7 +16,7 @@ private:
 	int m_skipUpdatesCounter;
 	QList<CPortal*> m_portals;
 
-	int maxBuffSize(void) const;
+	quint64 maxBuffSize(void) const;
 public:
 	explicit CDataTable(QObject *parent = 0);
